poly1305: Move software implementation of Poly1305State into a module
